+++ src/lightly_train/_methods/distillationv2/distillationv2_transform.py
@@ -21,7 +21,6 @@ from lightly_train._transforms.transform import (
     MethodTransformArgs,
     NormalizeArgs,
     RandomFlipArgs,
-    RandomResizeArgs,
     RandomResizedCropArgs,
     RandomRotationArgs,
     SolarizeArgs,
@@ -37,7 +36,7 @@ from lightly_train.types import (
 )
 
 
-class DistillationV2RandomResizeArgs(RandomResizeArgs):
+class DistillationV2RandomResizedCropArgs(RandomResizedCropArgs):
     min_scale: float = 0.14
 
 
@@ -60,8 +59,8 @@ class DistillationV2TransformArgs(MethodTransformArgs):
     image_size: ImageSizeTuple = (224, 224)
     channel_drop: ChannelDropArgs | None = None
     num_channels: int | Literal["auto"] = "auto"
-    random_resize: DistillationV2RandomResizeArgs | None = Field(
-        default_factory=DistillationV2RandomResizeArgs
+    random_resize: DistillationV2RandomResizedCropArgs | None = Field(
+        default_factory=DistillationV2RandomResizedCropArgs
     )
     random_flip: RandomFlipArgs | None = Field(default_factory=RandomFlipArgs)
     random_rotation: RandomRotationArgs | None = None
@@ -82,11 +81,9 @@ class DistillationV2Transform(MethodTransform):
 
         self.transform = ViewTransform(
             ViewTransformArgs(
+                image_size=transform_args.image_size,
                 channel_drop=transform_args.channel_drop,
-                random_resized_crop=RandomResizedCropArgs(
-                    size=transform_args.image_size,
-                    scale=transform_args.random_resize,
-                ),
+                random_resized_crop=transform_args.random_resize,
                 random_flip=transform_args.random_flip,
                 random_rotation=transform_args.random_rotation,
                 color_jitter=transform_args.color_jitter,