+++ src/lightly_train/_task_models/image_classification_multihead/transforms.py
@@ -20,14 +20,14 @@ from lightly_train._transforms.transform import (
     ColorJitterArgs,
     NormalizeArgs,
     RandomFlipArgs,
-    RandomResizeArgs,
+    RandomResizedCropArgs,
     RandomRotate90Args,
     RandomRotationArgs,
 )
 from lightly_train.types import ImageSizeTuple
 
 
-class ImageClassificationRandomResizeArgs(RandomResizeArgs):
+class ImageClassificationRandomResizedCropArgs(RandomResizedCropArgs):
     min_scale: float = 0.08
     max_scale: float = 1.0
 
@@ -41,8 +41,8 @@ class ImageClassificationMultiheadTrainTransformArgs(ImageClassificationTransfor
     channel_drop: ChannelDropArgs | None = None
     num_channels: int | Literal["auto"] = "auto"
     normalize: NormalizeArgs | Literal["auto"] = "auto"
-    random_crop: ImageClassificationRandomResizeArgs | None = Field(
-        default_factory=ImageClassificationRandomResizeArgs
+    random_crop: ImageClassificationRandomResizedCropArgs | None = Field(
+        default_factory=ImageClassificationRandomResizedCropArgs
     )
     resize_scale: float | None = None
     random_flip: RandomFlipArgs | None = Field(default_factory=RandomFlipArgs)
@@ -85,7 +85,7 @@ class ImageClassificationMultiheadValTransformArgs(ImageClassificationTransformA
     channel_drop: ChannelDropArgs | None = None
     num_channels: int | Literal["auto"] = "auto"
     normalize: NormalizeArgs | Literal["auto"] = "auto"
-    random_crop: ImageClassificationRandomResizeArgs | None = None
+    random_crop: ImageClassificationRandomResizedCropArgs | None = None
     resize_scale: float | None = 1.143
     random_flip: RandomFlipArgs | None = None
     random_rotate_90: RandomRotate90Args | None = None